More rendering work
Still doesn't work, but I have realised that having each physical object
being a object is not exactly the best for preformance, and causes
having as little as a 15*15 area full of simple objects that do little
simulation and no looping causes 14 or so minute startup times and 100%
command usage. Basically, I'll have to start all over in favor of a less
resource intensive setup.